Move `GasEstimator` package
