+++ src/common/blockchain/index.js
@@ -1,6 +1,7 @@
 import * as Ethereum from './ethereum'
 import * as Plasma from './plasma'
 import * as Contract from './contract'
+import * as GasEstimator from './gasEstimator'
 import * as TxOptions from './txOptions'
 import * as Token from './token'
 import * as Wait from './wait'
@@ -10,6 +11,7 @@ export {
   Ethereum,
   Plasma,
   Contract,
+  GasEstimator,
   TxOptions,
   BlockchainDataFormatter,
   Token,
