+++ src/common/utils/index.js
@@ -3,7 +3,6 @@ import * as Price from './price'
 import * as Storage from './storage'
 import * as Datetime from './datetime'
 import * as Formatter from './formatter'
-import * as GasEstimator from './gasEstimator'
 import * as Alerter from './alerter'
 import * as Identicon from './identicon'
 import * as Notification from './notification'
@@ -27,7 +26,6 @@ export {
   Datetime,
   Dimensions,
   Formatter,
-  GasEstimator,
   Mapper,
   Notification,
   OmgUtil,
