+++ src/components/views/exit/ExitForm.js
@@ -15,8 +15,8 @@ import {
   OMGFeeInput
 } from 'components/widgets'
 import { TransferHelper } from 'components/views/transfer'
-import { Validator, GasEstimator } from 'common/utils'
-import { Plasma } from 'common/blockchain'
+import { Validator } from 'common/utils'
+import { Plasma, GasEstimator } from 'common/blockchain'
 import { OMGBlockchainLabel } from 'components/widgets'
 import { ScrollView } from 'react-native-gesture-handler'
 import { paramsForTransferFormToTransferSelectFee } from 'components/views/transfer/transferNavigation'
